Support assigning labels to stack elements
Using a new `$<label>` syntax, which assigns a label to the last element
on stack. This can be specified on the same line as the push/opcode or
in a standalone line.

This is meant to make it easier to keep track of things while developing
and doesn't have an effect on the generated Script.

Depends on https://github.com/bit-matrix/wiz-data/pull/2 and
https://github.com/bit-matrix/script-wiz-lib-core/pull/5.